requirements.txt
python-telegram-bot==25.3
requests
pytz


